pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "pyoverleaf>=0.1.5",
    "typer>=0.16.0",
]

[project.scripts]
mcp = "mcp:main"
overleaf-cli = "mcp.overleaf_api.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
